refactor: 마이페이지 패키지 구조 개선
