+++ frontend/rush/src/components/myPage/Box.js
