+++ frontend/rush/src/components/myPage/CancelButton.js
