+++ frontend/rush/src/constants/ADDRESS.js
@@ -1,4 +1,4 @@
 export const BACKEND_ADDRESS = "http://seoultechfootprint.shop:8080";
 export const FRONTEND_ADDRESS = "http://seoultechfootprint.shop:80";
-//export const BACKEND_ADDRESS = "http://localhost:8080";
-//export const FRONTEND_ADDRESS = "http://localhost:3000";
+// export const BACKEND_ADDRESS = "http://localhost:8080";
+// export const FRONTEND_ADDRESS = "http://localhost:3000";
